+++ processing/schematisation_algorithms.py
@@ -1,0 +1,111 @@
+# -*- coding: utf-8 -*-
+
+"""
+***************************************************************************
+*                                                                         *
+*   This program is free software; you can redistribute it and/or modify  *
+*   it under the terms of the GNU General Public License as published by  *
+*   the Free Software Foundation; either version 2 of the License, or     *
+*   (at your option) any later version.                                   *
+*                                                                         *
+***************************************************************************
+"""
+
+import os
+import shutil
+from uuid import uuid4
+
+# from sqlalchemy.exc import OperationalError
+from threedi_modelchecker.threedi_database import ThreediDatabase
+# from threedi_modelchecker.model_checks import ThreediModelChecker
+from threedi_modelchecker.schema import ModelSchema
+from threedi_modelchecker import errors
+from qgis.PyQt.QtCore import QCoreApplication
+from qgis.core import (
+    QgsProcessingAlgorithm,
+    QgsProcessingParameterFile
+)
+
+
+def backup_sqlite(filename):
+    """Make a backup of the sqlite database."""
+    backup_sqlite_path = None
+    backup_folder = os.path.join(os.path.dirname(os.path.dirname(filename)), "_backup")
+    os.makedirs(backup_folder, exist_ok=True)
+    prefix = str(uuid4())[:8]
+    backup_sqlite_path = os.path.join(backup_folder, f"{prefix}_{os.path.basename(filename)}")
+    shutil.copyfile(filename, backup_sqlite_path)
+    return backup_sqlite_path
+
+
+class MigrateAlgorithm(QgsProcessingAlgorithm):
+    """
+    Migrate 3Di model schema to the current version
+    """
+    INPUT = 'INPUT'
+    OUTPUT = 'OUTPUT'
+
+    def initAlgorithm(self, config):
+        self.addParameter(
+            QgsProcessingParameterFile(
+                self.INPUT,
+                self.tr('3Di Spatialite'),
+                extension="sqlite"
+            )
+        )
+
+    def processAlgorithm(self, parameters, context, feedback):
+        success = False
+        filename = self.parameterAsFile(parameters, self.INPUT, context)
+        db_type = "spatialite"
+        db_settings = {"db_path": filename}
+        threedi_db = ThreediDatabase(db_settings, db_type=db_type)
+        schema = ModelSchema(threedi_db)
+        try:
+            schema.validate_schema()
+        except errors.MigrationMissingError:
+            backup_filepath = backup_sqlite(filename)
+            schema.upgrade(backup=False)
+            shutil.rmtree(os.path.dirname(backup_filepath))
+        success = True
+        return {self.OUTPUT: success}
+
+    def name(self):
+        """
+        Returns the algorithm name, used for identifying the algorithm. This
+        string should be fixed for the algorithm, and must not be localised.
+        The name should be unique within each provider. Names should contain
+        lowercase alphanumeric characters only and no spaces or other
+        formatting characters.
+        """
+        return 'migrate'
+
+    def displayName(self):
+        """
+        Returns the translated algorithm name, which should be used for any
+        user-visible display of the algorithm name.
+        """
+        return self.tr('Migrate Spatialite')
+
+    def group(self):
+        """
+        Returns the name of the group this algorithm belongs to. This string
+        should be localised.
+        """
+        return self.tr(self.groupId())
+
+    def groupId(self):
+        """
+        Returns the unique ID of the group this algorithm belongs to. This
+        string should be fixed for the algorithm, and must not be localised.
+        The group id should be unique within each provider. Group id should
+        contain lowercase alphanumeric characters only and no spaces or other
+        formatting characters.
+        """
+        return 'Schematisation'
+
+    def tr(self, string):
+        return QCoreApplication.translate('Processing', string)
+
+    def createInstance(self):
+        return MigrateAlgorithm()